pyproject.toml
[project]
name = "internal-research-pipeline"
version = "1.0.0"
description = "Automated pipeline to discover agentic AI papers in finance/audit and pharma/medicine from top ML conferences."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "requests>=2.31",
    "pandas>=2.1",
    "tqdm>=4.66",
    "openreview-py>=0.8",
    "pyyaml>=6.0",
]

[project.scripts]
research-pipeline = "research_pipeline.pipeline:main"
curate-results = "research_pipeline.curate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research_pipeline"]

[dependency-groups]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
]
